Move prim type icon provider into its own file for re-use
